+++ src/components/SitemapPage.tsx
@@ -4,7 +4,7 @@ import { completeSitemap, dynamicPaths } from '@/config/sitemap';
 import { Link } from 'react-router-dom';
 import { ChevronRight } from 'lucide-react';
 import { SEO } from './SEO';
-import { AppLayout } from './layout/AppLayout';
+import { AppLayout } from '@/layout/AppLayout';
 
 export const SitemapPage: React.FC = () => {
   return (
