+++ src/pages/ContractBuilder.tsx
@@ -1,7 +1,7 @@
 
 import React from 'react';
 import { SEO } from '@/components/SEO';
-import { AppHeader } from '@/components/layout/AppHeader';
+import { AppHeader } from '@/layout/AppHeader';
 import { Footer } from '@/components/Footer';
 import { ProtectedRoute } from '@/components/ProtectedRoute';
 import { ContractBuilderContent } from '@/components/contracts/page/ContractBuilderContent';
