+++ src/pages/ForumCategoryPage.tsx
@@ -4,7 +4,7 @@ import { useParams, Link } from "react-router-dom";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { SEO } from "@/components/SEO";
-import { AppLayout } from "@/components/layout/AppLayout";
+import { AppLayout } from "@/layout/AppLayout";
 import PostCard from "@/components/community/PostCard";
 import { ForumPost, ForumCategoryInfo } from "@/types/community";
 import { Badge } from "@/components/ui/badge";
