+++ src/pages/HiringTracker.tsx
@@ -1,7 +1,7 @@
 
 import { useState } from "react";
 import { useParams } from "react-router-dom";
-import { AppHeader } from "@/components/layout/AppHeader"; 
+import { AppHeader } from "@/layout/AppHeader"; 
 import { Footer } from "@/components/Footer";
 import { KanbanBoard } from "@/components/hiring-tracker/KanbanBoard";
 import { HiringAnalytics } from "@/components/hiring-tracker/HiringAnalytics";
