+++ src/pages/SearchPage.tsx
@@ -5,7 +5,7 @@ import { EnhancedSearchInput } from "@/components/search/EnhancedSearchInput";
 import { generateSearchSuggestions } from "@/data/marketplaceData";
 import { SearchSuggestion } from "@/types/search";
 import { useAISearch } from "@/hooks/useAISearch";
-import { AppLayout } from "@/components/layout/AppLayout";
+import { AppLayout } from "@/layout/AppLayout";
 
 export default function SearchPage() {
   const [params] = useSearchParams();
