+++ src/pages/admin/AppMetadataManager.tsx
@@ -1,6 +1,6 @@
 
 import React from "react";
-import { AppLayout } from "@/components/layout/AppLayout";
+import { AppLayout } from "@/layout/AppLayout";
 import { SEO } from "@/components/SEO";
 import { MetadataManager } from "@/components/mobile-app/admin/MetadataManager";
