+++ src/pages/dashboard/DisputeDetailPage.tsx
@@ -1,7 +1,7 @@
 
 import React from 'react';
 import { SEO } from '@/components/SEO';
-import { AppHeader } from '@/components/layout/AppHeader';
+import { AppHeader } from '@/layout/AppHeader';
 import { Footer } from '@/components/Footer';
 import { DisputeDetail } from '@/components/disputes';
 import { ProtectedRoute } from '@/components/ProtectedRoute';
